+++ src/test/java/io/anserini/encoder/dense/BgeBaseEn15EncoderInferenceTest.java
@@ -946,7 +946,7 @@ public class BgeBaseEn15EncoderInferenceTest extends DenseEncoderInferenceTest {
               -0.57727516f, 0.6823535f, 0.8443152f, -0.29511172f } }
   };
 
-  static protected final Object[][] LONG_EXAMPLES = new Object[][] {
+  static private final Object[][] LONG_EXAMPLES = new Object[][] {
       { new String[] {
           "In the dawn of the 21st century, humanity stands on the brink of one of the most transformative periods in history: the rise of artificial intelligence (AI). " +
               "This technological revolution promises to redefine the way we live, work, and interact with the world around us. " +
@@ -1133,4 +1133,52 @@ public class BgeBaseEn15EncoderInferenceTest extends DenseEncoderInferenceTest {
   // We're running into this issue on GitHub Java CI:
   // > Error: The operation was canceled.
   // Can't reproduce locally, but separating test cases into separate files seems to fix it...
+
+  @Test
+  public void testMaxLength() throws OrtException, IOException, URISyntaxException {
+    try(DenseEncoder encoder = new BgeBaseEn15Encoder()) {
+      float[] expectedWeights = (float[]) BgeBaseEn15EncoderInferenceTest.LONG_EXAMPLES[0][1];
+      String[] inputStrings = (String[]) BgeBaseEn15EncoderInferenceTest.LONG_EXAMPLES[0][0];
+
+      float[] outputs = encoder.encode(inputStrings[0]);
+      assertArrayEquals(expectedWeights, outputs, 1e-4f);
+    }
+
+    // Specify the directory for which you want to check the free space.
+    // You can use the current directory, a specific drive, or a subdirectory.
+    File directory = new File("."); // Current directory
+
+    // Get the free space in bytes.
+    long freeSpace = directory.getFreeSpace();
+
+    // Get the total space in bytes.
+    long totalSpace = directory.getTotalSpace();
+
+    // Get the usable space in bytes (space available to this JVM).
+    long usableSpace = directory.getUsableSpace();
+
+    // Convert bytes to gigabytes for easier reading.
+    double freeSpaceGB = (double) freeSpace / (1024 * 1024 * 1024);
+    double totalSpaceGB = (double) totalSpace / (1024 * 1024 * 1024);
+    double usableSpaceGB = (double) usableSpace / (1024 * 1024 * 1024);
+
+    // Print the results.
+    System.out.printf("Total space: %.2f GB%n", totalSpaceGB);
+    System.out.printf("Usable space: %.2f GB%n", usableSpaceGB);
+    System.out.printf("Free space: %.2f GB%n", freeSpaceGB);
+
+    Runtime runtime = Runtime.getRuntime();
+
+    long maxMemory = runtime.maxMemory(); //Maximum amount of memory that the JVM will attempt to use
+    long totalMemory = runtime.totalMemory(); // Total memory currently available to the JVM
+    long freeMemory = runtime.freeMemory();  // Amount of free memory available in the JVM
+
+    // Calculate used memory
+    long usedMemory = totalMemory - freeMemory;
+
+    System.out.println("Max memory: " + maxMemory / (1024 * 1024) + "MB");
+    System.out.println("Total memory: " + totalMemory / (1024 * 1024) + "MB");
+    System.out.println("Free memory: " + freeMemory / (1024 * 1024) + "MB");
+    System.out.println("Used memory: " + usedMemory / (1024 * 1024) + "MB");
+  }
 }
